+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.6.0]
 ### Added
 ### Changed
 - Adjusted ignore settings in default settings file to ignore hidden 
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Deprecated
 ### Removed
 ### Fixed
+- With every run of zettels, the index was built from scratch. Now it only 
+  updates it if files in the Zettelkasten directory have changed.
 ### Security
 
 ## [0.5.0] - 2017-06-09